Modularised the code
